Moved contactMap into shared code
The contact map is in ExtendedApplication now and all the Activities
that use it go through that.  As a result, the map is only created for
the first activity that uses it and persists for the others (until the
Application is destroyed).  So the app is faster and the code is
cleaner/less duplication.