Handle screenSize config change
When I upped the targetSdkVersion, rotating the screen while processing
would leak the activity.  I learned that you need to include screenSize
as a configChange you handle in addition to orientation on API 13+
(3.2+).  Probably because the navigation buttons could be software
buttons and they move on rotate so the dimensions change.