Fixes for the refresh icon on first load
The progress icon wasn't working on first load correctly, because
onCreate is executed before the menu is actually created.  The solution
fixes it most of the time but might still have a race condition causing
it to load with the progress icon still going even though the processing
thread is done.